Added tests for the debuggers, and various QoL APIs for the debugger.
