+++ src/Draco.Debugger.Tui/Draco.Debugger.Tui.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="Terminal.Gui" Version="1.11.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Properties\" />
+  </ItemGroup>
+
 </Project>
